Move rendering in Window singleton to give access to window settings
