+++ crates/modor_graphics/Cargo.toml
@@ -20,3 +20,6 @@ winit = "0.26"
 wgpu = "0.12"
 modor = { version = "0.1.0", path = "../modor" }
 modor_physics = { version = "0.1.0", path = "../modor_physics" }
+
+[dev-dependencies]
+rand = "0.8"
