+++ crates/modor_graphics/src/internal/mod.rs
@@ -1,5 +1,4 @@
 #[macro_use]
 mod utils;
 
-pub(crate) mod context;
 pub(crate) mod window;
