+++ crates/modor_graphics/src/lib.rs
@@ -3,14 +3,17 @@
 #[macro_use]
 extern crate modor;
 
+#[macro_use]
+mod internal;
 mod appearance;
 mod backend;
-mod internal;
+mod background;
 mod module;
 mod runner;
 mod window;
 
 pub use appearance::*;
+pub use background::*;
 pub use module::*;
 pub use runner::*;
 pub use window::*;
