+++ crates/modor_graphics/src/runner.rs
@@ -1,8 +1,9 @@
-use crate::internal::context::Context;
 use crate::internal::window::WindowInit;
 use modor::App;
+use std::sync::{Arc, RwLock, RwLockReadGuard};
 use winit::event::{Event, WindowEvent};
 use winit::event_loop::{ControlFlow, EventLoop};
+use winit::window::Window as WinitWindow;
 
 // TODO: move frame rate limitation in runner ?
 
@@ -15,19 +16,18 @@ pub fn runner(mut app: App) {
     let window =
         window.expect("failed to create window because `GraphicsModule` entity is not found");
     event_loop.run(move |event, _, control_flow| match event {
-        Event::MainEventsCleared => window.request_redraw(),
-        Event::RedrawRequested(window_id) if window_id == window.id() => app.update(),
-        Event::WindowEvent { event, window_id } if window_id == window.id() => match event {
-            WindowEvent::Resized(physical_size) => app.run_for_singleton::<Context, _>(|c| {
-                c.resize(physical_size.width, physical_size.height)
-            }),
-            WindowEvent::ScaleFactorChanged { new_inner_size, .. } => app
-                .run_for_singleton::<Context, _>(|c| {
-                    c.resize(new_inner_size.width, new_inner_size.height)
-                }),
-            WindowEvent::CloseRequested => *control_flow = ControlFlow::Exit,
-            _ => {}
-        },
+        Event::MainEventsCleared => read_window(&window).request_redraw(),
+        Event::RedrawRequested(window_id) if window_id == read_window(&window).id() => app.update(),
+        Event::WindowEvent { event, window_id } if window_id == read_window(&window).id() => {
+            match event {
+                WindowEvent::CloseRequested => *control_flow = ControlFlow::Exit,
+                _ => {}
+            }
+        }
         _ => {}
     });
+}
+
+fn read_window(window: &Arc<RwLock<WinitWindow>>) -> RwLockReadGuard<'_, WinitWindow> {
+    window.read().expect("internal error: cannot read window")
 }
